+++ lib/glorbo_web/live/proposals_live.ex
@@ -24,7 +24,7 @@ defmodule GlorboWeb.ProposalsLive do
   @impl true
   def mount(%{"company" => co}, _session, socket) do
     cond do
-      not GlorboWeb.Slug.valid?(co) ->
+      not Glorbo.Slug.valid?(co) ->
         {:ok,
          socket
          |> put_flash(:error, "Invalid company identifier.")