+++ lib/glorbo_web/live/skills_live.ex
@@ -27,7 +27,7 @@ defmodule GlorboWeb.SkillsLive do
   @impl true
   def mount(%{"company" => co}, _session, socket) do
     cond do
-      not GlorboWeb.Slug.valid?(co) ->
+      not Glorbo.Slug.valid?(co) ->
         {:ok,
          socket
          |> put_flash(:error, "Invalid company identifier.")